+++ deploy/docker/seqr/entrypoint.sh
@@ -25,7 +25,7 @@ if [ -e "/.config/service-account-key.json" ]; then
     until [ "$retries" -ge 5 ]
     do
         gcloud auth activate-service-account --key-file /.config/service-account-key.json && break
-        retries=$((retries+1)) 
+        retries=$((retries+1))
         echo "gcloud auth failed. Retrying, attempt ${retries}/5"
         sleep 10
     done
@@ -39,6 +39,21 @@ cd /seqr
 echo "*:*:*:*:$POSTGRES_PASSWORD" > ~/.pgpass
 chmod 600 ~/.pgpass
 cat ~/.pgpass
+
+# wait for database connectivity, exit if we don't get it within ~2 minutes
+pg_retries=0
+until [ "$pg_retries" -ge 10 ]
+do
+    pg_isready -d postgres -h "$POSTGRES_SERVICE_HOSTNAME" -U postgres && break
+    pg_retries=$((pg_retries+1))
+    if [ "$pg_retries" -eq 10 ]; then
+        echo "Postgres database wasn't available after 10 connection attempts"
+        exit 1
+    else
+        echo "Unable to connect to postgres, retrying. Attempt ${pg_retries}/10"
+        sleep 12
+    fi
+done
 
 # init and populate seqrdb unless it already exists
 if ! psql --host "$POSTGRES_SERVICE_HOSTNAME" -U postgres -l | grep seqrdb; then
@@ -64,13 +79,13 @@ if [ "$RUN_CRON_JOBS" ]; then
     # set up cron jobs
     # shellcheck disable=SC2016
     echo 'SHELL=/bin/bash
-0 0 * * 0 /usr/local/bin/python /seqr/manage.py update_omim --omim-key=$OMIM_KEY >> /proc/1/fd/1 2>&1
-0 0 * * 0 /usr/local/bin/python /seqr/manage.py update_human_phenotype_ontology >> /proc/1/fd/1 2>&1
-0 0 * * 0 /usr/local/bin/python /seqr/manage.py import_all_panels https://panelapp.agha.umccr.org/api/v1 --label=AU >> /proc/1/fd/1 2>&1
-0 0 * * 0 /usr/local/bin/python /seqr/manage.py import_all_panels https://panelapp.genomicsengland.co.uk/api/v1 --label=UK >> /proc/1/fd/1 2>&1
-0 12 * * 1 /usr/local/bin/python /seqr/manage.py detect_inactive_privileged_users >> /proc/1/fd/1 2>&1
-0 2 * * * /usr/local/bin/python /seqr/manage.py check_bam_cram_paths >> /proc/1/fd/1 2>&1
-0 0 1 1/3 * /usr/local/bin/python /seqr/manage.py clear_project_tags ALL_USER_DEMO --skip-confirm >> /proc/1/fd/1 2>&1
+0 0 * * 0 /opt/venv/bin/python /seqr/manage.py update_omim --omim-key=$OMIM_KEY >> /proc/1/fd/1 2>&1
+0 0 * * 0 /opt/venv/bin/python /seqr/manage.py update_human_phenotype_ontology >> /proc/1/fd/1 2>&1
+0 0 * * 0 /opt/venv/bin/python /seqr/manage.py import_all_panels https://panelapp.agha.umccr.org/api/v1 --label=AU >> /proc/1/fd/1 2>&1
+0 0 * * 0 /opt/venv/bin/python /seqr/manage.py import_all_panels https://panelapp.genomicsengland.co.uk/api/v1 --label=UK >> /proc/1/fd/1 2>&1
+0 12 * * 1 /opt/venv/bin/python /seqr/manage.py detect_inactive_privileged_users >> /proc/1/fd/1 2>&1
+0 2 * * * /opt/venv/bin/python /seqr/manage.py check_bam_cram_paths >> /proc/1/fd/1 2>&1
+0 0 1 1,4,7,10 * /opt/venv/bin/python /seqr/manage.py clear_project_tags ALL_USER_DEMO --skip-confirm >> /proc/1/fd/1 2>&1
 ' | crontab -
 
     env > /etc/environment  # this is necessary for crontab commands to run with the right env. vars.