+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## dev
 
+## 3/18/22
+* Disallow deleting individuals in matchmaker (REQUIRES DB MIGRATION)
+* Add GenCC reference data (REQUIRES DB MIGRATION)
+* Update local pipeline to use latest clinvar
+
 ## 3/10/22
 * Return hom alt SNPs in trans with Deletions as compound hets
 * Disable variant download in demo projects (REQUIRES DB MIGRATION)